backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
PyYAML==6.0.2

